Conflicts:
	.travis.yml
	cocos2dx/proj.emscripten/cocos2dx.mk
	cocos2dx/proj.nacl/cocos2dx.mk
	samples/Lua/TestLua/proj.linux/Makefile
	samples/Lua/TestLua/proj.nacl/Makefile
	scripting/lua/cocos2dx_support/CCLuaStack.h
	scripting/lua/cocos2dx_support/LuaCocos2d.cpp
	scripting/lua/proj.nacl/Makefile
	template/multi-platform-lua/proj.linux/Makefile

+++ cocos2dx/platform/platform.h
@@ -43,7 +43,7 @@ struct CC_DLL cc_timeval
 #else
     long    tv_sec;        // seconds
 #endif
-    long    tv_usec;    // microSeconds
+    int tv_usec;    // microSeconds
 };
 
 class CC_DLL CCTime
